+++ sphinx/domains/python/__init__.py
@@ -38,12 +38,15 @@ if TYPE_CHECKING:
 
 # re-export objects for backwards compatibility
 # xref https://github.com/sphinx-doc/sphinx/issues/12295
+from sphinx.domains.python._annotations import (  # NoQA: F401
+    _parse_arglist,  # for sphinx-immaterial
+    type_to_xref,
+)
 from sphinx.domains.python._object import (  # NoQA: F401
     PyField,
     PyGroupedField,
     PyTypedField,
     PyXrefMixin,
-    py_sig_re,
 )
 
 logger = logging.getLogger(__name__)